MP-89 :truck: Rename: 엔터티 유틸리티 클래스 폴더 이동
* 도메인 유틸리티 클래스와 분리하기 위한 폴더 이동
