+++ src/test/java/kr/modusplant/global/persistence/repository/SiteMemberJpaRepositoryTest.java
@@ -2,7 +2,7 @@ package kr.modusplant.global.persistence.repository;
 
 import kr.modusplant.global.persistence.entity.SiteMemberEntity;
 import kr.modusplant.support.context.RepositoryOnlyContext;
-import kr.modusplant.support.util.SiteMemberEntityTestUtils;
+import kr.modusplant.support.util.entity.SiteMemberEntityTestUtils;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
